feat(scheduler): add shard disk repair manager
1. finish shard task process and shard disk repair

with: #22452167

+++ blobstore/scheduler/disk_droper.go
@@ -423,7 +423,7 @@ func (mgr *DiskDropMgr) listUnMigratedVuid(ctx context.Context, diskID proto.Dis
 
 func (mgr *DiskDropMgr) initOneTask(ctx context.Context, src proto.Vuid, dropDiskID proto.DiskID, diskIDC string) {
 	t := proto.MigrateTask{
-		TaskID:       client.GenMigrateTaskID(proto.TaskTypeDiskDrop, dropDiskID, src.Vid()),
+		TaskID:       client.GenMigrateTaskID(proto.TaskTypeDiskDrop, dropDiskID, uint32(src.Vid())),
 		TaskType:     proto.TaskTypeDiskDrop,
 		State:        proto.MigrateStateInited,
 		SourceDiskID: dropDiskID,
